Merging updates to TF2 from rodrigo2019/keras_yolo2
This copies over hopefully all relevant files from the repository
of rodrigo2019 in his fork of experiencor/keras-yolo2, which this
port of YOLOv2 to ROS seems to based on as well.

The files have just been copied and not been changed at all as of
yet. This is a WIP.