Updating yolo_server.py script
This modifies the behaviour of yolo_server.py to be using the
utils.py function exactly as it is in the repository somewhat
upstream repo keras_yolo2 by experiencor and rodrigo2019.
Hopefully by not modifying utils.py directly in this downstream
repo keeping it in sync with upstream will be easier in the future.

Specifically it imports the additional functions introduces to
utils.py by datadigger, but moved into yolo_server.py instead of
utils.py. Namely these functions are get_all_labels, get_xy_score
get_xy_center and get_xy_extents.