Fixing prediction in ROS context
The way the data about a detected object is returned seems to be
different in the version of rodrigo2019 compared to experiencor.
This seems to be working for now. The way it is implemented tries
to leave everything in the code of the upstream package the same
as to make a future update of the upstream code easier.